SyncProcessor as interface
