+++ rskj-core/src/test/java/co/rsk/net/messages/MessageVisitorTest.java
@@ -5,7 +5,6 @@ import co.rsk.crypto.Keccak256;
 import co.rsk.net.*;
 import co.rsk.scoring.EventType;
 import co.rsk.scoring.PeerScoringManager;
-import co.rsk.validators.BlockValidationRule;
 import org.ethereum.config.Constants;
 import org.ethereum.core.Block;
 import org.ethereum.core.Transaction;
@@ -31,7 +30,7 @@ public class MessageVisitorTest {
     private ChannelManager channelManager;
     private PeerScoringManager peerScoringManager;
     private TransactionGateway transactionGateway;
-    private SyncProcessor syncProcessor;
+    private SyncProcessorImpl syncProcessor;
     private BlockProcessor blockProcessor;
     private RskSystemProperties config;
 
@@ -39,7 +38,7 @@ public class MessageVisitorTest {
     public void setUp() {
         config = mock(RskSystemProperties.class);
         blockProcessor = mock(BlockProcessor.class);
-        syncProcessor = mock(SyncProcessor.class);
+        syncProcessor = mock(SyncProcessorImpl.class);
         transactionGateway = mock(TransactionGateway.class);
         peerScoringManager = mock(PeerScoringManager.class);
         channelManager = mock(ChannelManager.class);
