+++ rskj-core/src/test/java/co/rsk/net/simples/SimpleAsyncNode.java
@@ -51,7 +51,7 @@ public class SimpleAsyncNode extends SimpleNode {
     private static final TestSystemProperties config = new TestSystemProperties();
     private ExecutorService executor = Executors.newSingleThreadExecutor();
     private LinkedBlockingQueue<Future> futures = new LinkedBlockingQueue<>(5000);
-    private SyncProcessor syncProcessor;
+    private SyncProcessorImpl syncProcessor;
     private SimpleChannelManager simpleChannelManager;
 
     public SimpleAsyncNode(MessageHandler handler, Blockchain blockchain) {
@@ -60,7 +60,7 @@ public class SimpleAsyncNode extends SimpleNode {
 
     public SimpleAsyncNode(MessageHandler handler,
                            Blockchain blockchain,
-                           SyncProcessor syncProcessor,
+                           SyncProcessorImpl syncProcessor,
                            SimpleChannelManager simpleChannelManager) {
         super(handler, blockchain);
         this.syncProcessor = syncProcessor;
@@ -110,7 +110,7 @@ public class SimpleAsyncNode extends SimpleNode {
         this.futures.clear();
     }
 
-    public SyncProcessor getSyncProcessor() {
+    public SyncProcessorImpl getSyncProcessor() {
         return this.syncProcessor;
     }
 
@@ -136,7 +136,7 @@ public class SimpleAsyncNode extends SimpleNode {
         PeerScoringManager peerScoringManager = RskMockFactory.getPeerScoringManager();
         SimpleChannelManager channelManager = new SimpleChannelManager();
         BlockFactory blockFactory = new BlockFactory(config.getActivationConfig());
-        SyncProcessor syncProcessor = new SyncProcessor(
+        SyncProcessorImpl syncProcessor = new SyncProcessorImpl(
                 blockchain, indexedBlockStore, mock(ConsensusValidationMainchainView.class), blockSyncService, syncConfiguration, blockFactory,
                 blockValidationRule,
                 new BlockCompositeRule(new BlockUnclesHashValidationRule(), new BlockRootValidationRule(config.getActivationConfig())),
